[onert/acl_cl] Refactor operation handling to use separate operation files
This commit refactors ACL-CL backend to move operation implementations from KernelGenerator.cc and Validator.h to separate operation files in ops/ directory.
This improves code organization and consistency with other backend patterns.

ONE-DCO-1.0-Signed-off-by: Hyeongseok Oh <hseok82.oh@samsung.com>
