+++ runtime/onert/backend/acl_cl/ops/PReLU.cc
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2025 Samsung Electronics Co., Ltd. All Rights Reserved
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "../KernelGenerator.h"
+#include "../Validator.h"
+
+#include <AclKernelGen.h>
+
+namespace onert::backend::acl_cl
+{
+
+void Validator::visit(const ir::operation::PReLU &) { _supported = true; }
+
+void KernelGenerator::visit(const ir::operation::PReLU &node)
+{
+  const auto ofm_index{node.getOutputs().at(0)};
+  const auto ifm_index{node.getInputs().at(ir::operation::PReLU::Input::INPUT)};
+  const auto alpha_index{node.getInputs().at(ir::operation::PReLU::Input::ALPHA)};
+
+  auto ofm_tensor = _tensor_reg->getAclTensor(ofm_index);
+  auto ifm_tensor = _tensor_reg->getAclTensor(ifm_index);
+  auto alpha_tensor = _tensor_reg->getAclTensor(alpha_index);
+
+  auto fn = acl_common::generateLayer<arm_compute::CLPReluLayer>(
+    ifm_tensor->handle(), alpha_tensor->handle(), ofm_tensor->handle());
+
+  _return_fn = acl_common::asAclFunction(std::move(fn));
+}
+
+} // namespace onert::backend::acl_cl
